mythmiscutil: move functions to mythcommandlineparser
