+++ mythtv/libs/libmythbase/test/test_mythcommandlineparser/test_mythcommandlineparser.cpp
@@ -192,4 +192,79 @@ void TestCommandLineParser::test_override_file (void)
     QCOMPARE(overrides["plover"], QString("plugh"));
 }
 
+void TestCommandLineParser::test_parse_cmdline_data(void)
+{
+    QTest::addColumn<QString>("input");
+    QTest::addColumn<QStringList>("expectedOutput");
+
+    QTest::newRow("simple")
+        << R"(This is a test string)"
+        << QStringList({"This", "is", "a", "test", "string"});
+    QTest::newRow("simplequotes")
+        << R"(cmd "whatever" "goes" here)"
+        << QStringList({R"(cmd)",
+                        R"("whatever")",
+                        R"("goes")",
+                        R"(here)"});
+    QTest::newRow("multiword")
+        << R"(cmd "whatever" "multi-word argument" arg3)"
+        << QStringList({R"(cmd)",
+                        R"("whatever")",
+                        R"("multi-word argument")",
+                        R"(arg3)"});
+    QTest::newRow("mixedargs")
+        << R"(cmd --arg1="whatever" --arg2="multi-word argument" --arg3)"
+        << QStringList({R"(cmd)",
+                        R"(--arg1="whatever")",
+                        R"(--arg2="multi-word argument")",
+                        R"(--arg3)"});
+    QTest::newRow("mixedquotes")
+        << R"(cmd --arg1 first-value --arg2 "second 'value'")"
+        << QStringList({R"(cmd)",
+                        R"(--arg1)",
+                        R"(first-value)",
+                        R"(--arg2)",
+                        R"("second 'value'")"});
+    QTest::newRow("mixeduneven")
+        << R"(cmd --arg1 first-value --arg2 "second 'value")"
+        << QStringList({R"(cmd)",
+                        R"(--arg1)",
+                        R"(first-value)",
+                        R"(--arg2)",
+                        R"("second 'value")"});
+    QTest::newRow("1escapedquote")
+        << R"(cmd -d --arg1 first-value --arg2 \"second)"
+        << QStringList({R"(cmd)",
+                        R"(-d)",
+                        R"(--arg1)",
+                        R"(first-value)",
+                        R"(--arg2)",
+                        R"(\"second)"});
+    QTest::newRow("nestedquotes")
+        << R"(cmd --arg1 first-value --arg2 "second \"value\"")"
+        << QStringList({R"(cmd)",
+                        R"(--arg1)",
+                        R"(first-value)",
+                        R"(--arg2)",
+                        R"("second \"value\"")"});
+    QTest::newRow("unfinishedquote")
+        << R"(cmd --arg1 first-value --arg2 "second \"value\")"
+        << QStringList({R"(cmd)",
+                        R"(--arg1)",
+                        R"(first-value)",
+                        R"(--arg2)",
+                        R"("second \"value\")"});
+}
+
+void TestCommandLineParser::test_parse_cmdline(void)
+{
+    QFETCH(QString, input);
+    QFETCH(QStringList, expectedOutput);
+
+    QStringList output = MythCommandLineParser::MythSplitCommandString(input);
+//  std::cerr << "Expected: " << qPrintable(expectedOutput.join("|")) << std::endl;
+//  std::cerr << "Actual:   " << qPrintable(output.join("|")) << std::endl;
+    QCOMPARE(output, expectedOutput);
+}
+
 QTEST_APPLESS_MAIN(TestCommandLineParser)
