--- mythtv/libs/libmythbase/test/test_mythmiscutil/.gitignore
@@ -1,1 +1,0 @@
-test_mythmiscutil
